added a parametric version of the prop
This is a reworked SolidWorks model that basically creates the same geometry as the previous prop-half model, but is more clean, clear, and adjustable. It has global variables to directly adjust parameters of interest: chord length at root and tip, thickness at root and tip, pitch angle at root and tip, diameter of root and tip, and minimum radius of the edge (for 3D printer compatibility).